cloud_function/requirements.txt
google-cloud-firestore==1.6.2
google-cloud-bigquery==1.24.0
google-api-python-client==1.7.11
google-cloud-pubsub==1.0.2